python_tasks/requirements.txt
logging==0.4.9.6
psycopg2-binary==2.9.10
